+++ tests/test_delta.py
@@ -80,24 +80,6 @@ def test_delta_as_int():
     assert int(delta) == int(secs)
 
 
-@parametrize('delta,expected', [
-    (Delta(seconds=742610.1512), (('weeks', 1),
-                                  ('days', 1),
-                                  ('hours', 14),
-                                  ('minutes', 16),
-                                  ('seconds', 50),
-                                  ('microseconds', 151200))),
-    (Delta(seconds=-742610.1512), (('weeks', -1),
-                                   ('days', -1),
-                                   ('hours', -14),
-                                   ('minutes', -16),
-                                   ('seconds', -50),
-                                   ('microseconds', -151200))),
-])
-def test_delta_as_iter(delta, expected):
-    assert tuple(delta) == expected
-
-
 @parametrize('obj,exception', [
     ({}, TypeError),
     ('', ParseError),